Improving dotenv implementation

+++ manager_lib/installer.py
@@ -58,15 +58,17 @@ def manager_init():
             cursor.execute("CREATE TABLE 'ADDITIONALPARAMS'(KEYID INT NOT NULL,PARAMETR TEXT NOT NULL,VALUE TEXT)")
     
             #Create .env
-            with open(f"{SCRIPT_HOME}/.env", "w") as env:
-                env.write(f"SCRIPT_HOME='{SCRIPT_HOME}' \nSSH_CONFIG='{USER_HOME}/.ssh/config' \nDATABASE_PATH='{SCRIPT_HOME}/ssh-conf.db'")
+            if not os.path.exists(f"{SCRIPT_HOME}/.env"):
+                with open(f"{SCRIPT_HOME}/.env", "w") as env:
+                    env.write(f"SCRIPT_HOME='{SCRIPT_HOME}' \nSSH_CONFIG='{USER_HOME}/.ssh/config' \nDATABASE_PATH='{SCRIPT_HOME}/ssh-conf.db'")
     else:
         #Create script home
         print(f"Creating script directory in {SCRIPT_HOME}")
         Path(SCRIPT_HOME).mkdir(parents=True, exist_ok=True)
         #Create .env
-        with open(f"{SCRIPT_HOME}/.env", "w") as env:
-            env.write(f"SCRIPT_HOME='{SCRIPT_HOME}' \nSSH_CONFIG='{USER_HOME}/.ssh/config' \nDATABASE_PATH='{SCRIPT_HOME}/ssh-conf.db'")
+        if not os.path.exists(f"{SCRIPT_HOME}/.env"):
+            with open(f"{SCRIPT_HOME}/.env", "w") as env:
+                env.write(f"SCRIPT_HOME='{SCRIPT_HOME}' \nSSH_CONFIG='{USER_HOME}/.ssh/config' \nDATABASE_PATH='{SCRIPT_HOME}/ssh-conf.db'")
     
         conn = sqlite3.connect(f'{SCRIPT_HOME}/ssh-conf.db')
         cursor = conn.cursor()
